Move `repos.RepoResult` -> `pkg.ScorecardResult`
